Improved token usage on chat init + Removed unnecesary prompting on image description + Minor refactoring
