+++ server/config/constants.py
@@ -8,3 +8,6 @@ CONTEXT_DOCUMENTS_FILE = "context_document_links.json"
 PROMPT_LIST_FILE = "prompt_list.json"
 
 TEMP_UPLOADS_PATH = Path('temp/uploads')
+
+FULL_DESC_TRIGGER = '###DESCRIBE_LONG_IMAGE###'
+SHORT_DESC_TRIGGER = '###DESCRIBE_SHORT_IMAGE###'
